@@ -1,28 +1,25 @@
 <?php
 
-use Salamek\Zasilkovna\Api;
+declare(strict_types=1);
 
-/**
- * Copyright (C) 2016 Adam Schubert <adam.schubert@sg1-game.net>.
- */
+
+use Salamek\Zasilkovna\IApi;
+
 abstract class BaseTest extends PHPUnit_Framework_TestCase
 {
-    /** @var null|Api */
-    protected $zasilkovnaApi = null;
+	protected ?IApi $zasilkovnaApi = null;
 
-    public function setUp()
-    {
-        $configPath = __DIR__ . '/config.json';
-        if (file_exists($configPath)) {
-            $config = json_decode(file_get_contents($configPath));
-            $this->zasilkovnaApi = new Api(
-                $config->apiPassword,
-                $config->apiKey
-            );
-        }
-        else
-        {
-            throw new \Exception('config.json not found');
-        }
-    }
+
+	public function setUp(): void
+	{
+		$configPath = __DIR__ . '/config.json';
+		if (file_exists($configPath)) {
+			$config = json_decode(file_get_contents($configPath));
+			$this->zasilkovnaApi = new \Salamek\Zasilkovna\ApiRest(
+				$config->apiPassword
+			);
+		} else {
+			throw new \RuntimeException('config.json not found');
+		}
+	}
 }
