@@ -1,15 +1,10 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: sadam
- * Date: 7.9.17
- * Time: 3:06
- */
+
+declare(strict_types=1);
 
 namespace Salamek\Zasilkovna\Exception;
 
 
-class RestFault extends \Exception
+final class RestFault extends \Exception
 {
-
 }
